agents/search-bot/requirements.txt
# Core ADK
google-adk>=2.1.0

# Google Cloud
google-cloud-aiplatform>=1.38.0
google-cloud-secretmanager>=2.16.0

# Common utilities
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Add your specific dependencies below
